+++ src/Clients/GuzzleClient.php
@@ -9,16 +9,18 @@ class GuzzleClient extends Client
 {
     /**
      * @param HetznerAPIClient $client
+     * @param array $additionalGuzzleConfig
      */
-    public function __construct(HetznerAPIClient $client)
+    public function __construct(HetznerAPIClient $client, $additionalGuzzleConfig = [])
     {
-        parent::__construct([
+        $guzzleConfig = array_merge([
             'base_uri' => $client->getBaseUrl(),
             'headers' => [
                 'Authorization' => 'Bearer '.$client->getApiToken(),
                 'Content-Type' => 'application/json',
                 'User-Agent' => ((strlen($client->getUserAgent()) > 0) ? $client->getUserAgent().' ' : '').'hcloud-php/'.HetznerAPIClient::VERSION,
             ],
-        ]);
+        ], $additionalGuzzleConfig);
+        parent::__construct($guzzleConfig);
     }
 }